+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './common';
 export * from './create-async';
 export * from './use-async';
-export * from './share-loading';
+export * from './share-pending';
 export * from './take-latest-promise';